requirements.txt
matplotlib
networkx
folium
numpy
Pillow
tensorflow
scikit-learn
